css/css-values/lh-rlh-on-root-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Values and Units Test: using lh and rlh units on the root element</title>
<link rel="author" title="Florian Rivoal" href="https://florian.rivoal.net/">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#font-relative-lengths">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  function get_root_font_size() {
    return window.getComputedStyle(window.document.documentElement).font_size;
  }
  function get_root_line_height() {
    return window.getComputedStyle(window.document.documentElement).lineHeight;
  }

  window.document.documentElement.style="font-size: initial; line-height:initial;";
  initial_f_s = get_root_font_size();
  initial_l_h = get_root_line_height()

  test(function() {
    window.document.documentElement.style="font-size: 42px; line-height: 1lh;";
    l_h = get_root_line_height();
    assert_equals( l_h, initial_l_h, "the lh unit on the root element's line-height property uses font metrics corresponding to the initial values of the font or line-height properties");
  }, "lh in line-height on root");

  test(function() {
    window.document.documentElement.style="font-size: 42px; line-height: 1rlh;";
    l_h = get_root_line_height();
    assert_equals( l_h, initial_l_h, "the rlh unit on the root element's line-height property uses font metrics corresponding to the initial values of the font or line-height properties");
  }, "rlh in line-height on root");

  test(function() {
    window.document.documentElement.style="font-size: 1lh; line-height: 42px;";
    f_s = get_root_font_size();
    assert_equals( f_s, initial_f_s, "the lh unit on the root element's font-size property uses font metrics corresponding to the initial values of the font or line-height properties");
  }, "lh in font-size on root");

  test(function() {
    window.document.documentElement.style="font-size: 1rlh; line-height: 42px;";
    f_s = get_root_font_size();
    assert_equals( f_s, initial_f_s, "the rlh unit on the root element's font-size property uses font metrics corresponding to the initial values of the font or line-height properties");

  }, "rlh in font-size on root");
</script>
